feat: refactor process tracking initialization
