+++ pulsar/src/pulsard/daemon.rs
@@ -1,10 +1,7 @@
 use std::collections::HashMap;
 
 use anyhow::{bail, Result};
-use bpf_common::{
-    log_error,
-    program::{BpfContext, Pinning, PERF_PAGES_DEFAULT},
-};
+use bpf_common::program::{BpfContext, Pinning, PERF_PAGES_DEFAULT};
 
 use pulsar_core::{
     bus::Bus,
@@ -58,13 +55,7 @@ impl PulsarDaemon {
         };
 
         let mut m = HashMap::new();
-        let process_tracker = match ProcessTrackerHandle::load_procfs() {
-            Ok(process_tracker) => process_tracker,
-            Err(err) => {
-                log_error("error initializing process tracker from procfs", err);
-                ProcessTrackerHandle::new()
-            }
-        };
+        let process_tracker = ProcessTrackerHandle::new();
 
         let general_config = config.get_module_config(GENERAL_CONFIG).unwrap_or_default();
         let perf_pages = general_config.with_default("perf_pages", PERF_PAGES_DEFAULT)?;
